part2/requirements.txt
flask
flask-restx
flask-bcrypt
flask-jwt-extended
sqlalchemy
flask-sqlalchemy
werkzeug.exceptions
flask-migrate
